backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.4
